requirements.txt
django-post-office<3.5
django<3.2.0
python-decouple<3.4
requests<2.25.0